address review comments and cleanup
