+++ src/functional/fencil_processors/codegens/gtfn/gtfn_ir.py
@@ -12,7 +12,6 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-
 from typing import ClassVar, List, Union
 
 import eve
