+++ src/functional/iterator/__init__.py
@@ -1,12 +1,9 @@
 from typing import Optional, Union
 
-from . import builtins, runtime, tracing
-
-
-__all__ = ["builtins", "runtime", "tracing"]
-
 from packaging.version import LegacyVersion, Version, parse
 from pkg_resources import DistributionNotFound, get_distribution  # type: ignore
+
+from . import tracing  # noqa: F401 # ignore unused
 
 
 try:
